Refactor meal-related classes and introduce MealIngredient functionality with new DTOs and repository

+++ src/main/java/org/tcpmanager/tcpmanager/calories/dto/IngredientResponse.java
@@ -1,5 +1,7 @@
 package org.tcpmanager.tcpmanager.calories.dto;
 
-public record IngredientResponse(Long id, String name, int calories, int fat, int carbs, int protein, String ean) {
+import java.math.BigDecimal;
+
+public record IngredientResponse(Long id, String name, BigDecimal calories, BigDecimal fat, BigDecimal carbs, BigDecimal protein, String ean) {
 
 }
